Split out isInstallingTheme
