+++ client/state/themes/selectors/index.js
@@ -38,26 +38,7 @@ import 'state/themes/init';
 
 export { getTheme } from 'state/themes/selectors/get-theme';
 export { getCanonicalTheme } from 'state/themes/selectors/get-canonical-theme';
-
-/**
- * When wpcom themes are installed on Jetpack sites, the
- * theme id is suffixed with -wpcom. Some operations require
- * the use of this suffixed ID. This util function adds the
- * suffix if the site is jetpack and the theme is not yet
- * installed on the site.
- *
- * @param {object} state	Global state tree
- * @param {string} themeId	Theme ID
- * @param {number} siteId	Site ID
- * @returns {string} 		Potentially suffixed theme ID
- */
-const getSuffixedThemeId = ( state, themeId, siteId ) => {
-	const siteIsJetpack = siteId && isJetpackSite( state, siteId );
-	if ( siteIsJetpack && ! getTheme( state, siteId, themeId ) ) {
-		return `${ themeId }-wpcom`;
-	}
-	return themeId;
-};
+export { isInstallingTheme } from 'state/themes/selectors/is-installing-theme';
 
 /**
  * Returns theme request error object
@@ -547,19 +528,6 @@ export function isActivatingTheme( state, siteId ) {
  */
 export function hasActivatedTheme( state, siteId ) {
 	return get( state.themes.completedActivationRequests, siteId, false );
-}
-
-/**
- * Whether the theme is currently being installed on the (Jetpack) site.
- *
- * @param  {object}  state   Global state tree
- * @param  {string}  themeId Theme ID for which we check installing state
- * @param  {number}  siteId  Site ID
- * @returns {boolean}         True if theme installation is ongoing
- */
-export function isInstallingTheme( state, themeId, siteId ) {
-	const suffixedThemeId = getSuffixedThemeId( state, themeId, siteId );
-	return get( state.themes.themeInstalls, [ siteId, suffixedThemeId ], false );
 }
 
 /**
